[Windows] Set the new text message before trying to set the icon
This fixes an issue in the code where we don't want to reload the same
icon over and over again, so we're stopping the function execution.

But that also stopped the message from being updated. So move the message
setting to the beginning so it's always executed.

Bug 36546 - [Windows Toolbar] On closing the solution XS status bar does
not update,it still shows Warnings or Error